+++ src/relativePath.js
@@ -1,0 +1,4 @@
+import * as Path from 'path';
+
+// Separated into its own file for jest mocking
+export const relativePath = filename => Path.relative('./', Path.resolve('./', filename));